Implement command calls (bpf-profile #11)
