+++ src/gen/tests/mod.rs
@@ -9,22 +9,6 @@ use std::io::Cursor;
 use std::path::Path;
 
 #[test]
-fn header_missing() {
-    let reader = Cursor::new(b"Lorem ipsum dolor sit amet");
-    let r = trace::contains_standard_header(reader);
-    assert!(r.is_ok());
-    assert!(!r.unwrap());
-}
-
-#[test]
-fn header_ok() {
-    let reader = Cursor::new(mock::SIMPLE_INPUT);
-    let r = trace::contains_standard_header(reader);
-    assert!(r.is_ok());
-    assert!(r.unwrap());
-}
-
-#[test]
 fn generate_integral() {
     let resolv = resolver::Resolver::default();
     let reader = Cursor::new(mock::SIMPLE_INPUT);
