+++ src/main.rs
@@ -16,12 +16,18 @@
 #![forbid(unsafe_code)]
 #![deny(warnings)]
 
-mod buf;
+mod bpf;
+mod calls;
 mod cli;
 mod config;
 mod error;
+mod filebuf;
 mod gen;
+mod resolver;
 mod trace;
+
+#[cfg(test)]
+mod tests;
 
 fn main() {
     init_logger();
@@ -44,6 +50,14 @@ use crate::error::Result;
 /// Dispatches CLI commands.
 fn execute(app: cli::Application) -> Result<()> {
     match app.cmd {
+        cli::Command::Calls { trace, dump, tab } => {
+            calls::run(
+                &trace,
+                dump.as_ref().map(|p| p.as_ref()), // Option<T> -> Option<&T>,
+                tab,
+            )?;
+        }
+
         cli::Command::Generate {
             trace,
             asm,
@@ -58,9 +72,6 @@ fn execute(app: cli::Application) -> Result<()> {
                 &format,
                 output.as_ref().map(|p| p.as_ref()), // Option<T> -> Option<&T>
             )?;
-        }
-        cli::Command::Trace { trace } => {
-            trace::run(&trace)?;
         }
     }
